Revert "ksud: fork daemon to run late-load mode (https://github.com/tiann/KernelSU/pull/3334)"
This reverts commit 690042684b17064f228278d35f51b54f5b6d0224.
