+++ userspace/ksud/src/init_event.rs
@@ -1,15 +1,17 @@
 use crate::module::{handle_updated_modules, prune_modules};
-use crate::utils::{is_safe_mode, switch_mnt_ns};
+use crate::utils::{is_safe_mode, switch_cgroups, switch_mnt_ns};
 use crate::{
     assets, defs, ksucalls, metamodule, restorecon,
     utils::{self},
 };
-use anyhow::{Context, Result};
-use libc::_exit;
+use anyhow::{Context, Result, bail};
+use libc::{_exit, waitpid};
 use log::{info, warn};
 use prop_rs_android::resetprop::ResetProp;
 use prop_rs_android::sys_prop;
-use rustix::process::chdir;
+use rustix::fs::{Mode, OFlags, open};
+use rustix::process::{chdir, setpgid};
+use rustix::stdio::{dup2_stderr, dup2_stdin, dup2_stdout};
 use std::path::Path;
 use std::process::Command;
 
@@ -223,11 +225,43 @@ fn catch_bootlog(logname: &str, command: &[&str]) -> Result<()> {
 }
 
 pub fn soft_reboot() -> Result<()> {
-    utils::daemonize(|| -> Result<()> {
-        switch_mnt_ns(1)?;
-        chdir("/")?;
-        Ok(())
-    })?;
+    unsafe {
+        let pid = libc::fork();
+        if pid < 0 {
+            bail!("fork error {}", std::io::Error::last_os_error());
+        } else if pid > 0 {
+            loop {
+                if waitpid(pid, std::ptr::null_mut(), 0) < 0 {
+                    if *libc::__errno() != libc::EINTR {
+                        _exit(1);
+                    }
+                } else {
+                    break;
+                }
+            }
+            _exit(0);
+        }
+    }
+
+    setpgid(None, None)?;
+    switch_cgroups();
+    switch_mnt_ns(1)?;
+    chdir("/")?;
+    {
+        let null_fd = open("/dev/null", OFlags::RDWR, Mode::empty())?;
+        dup2_stdin(&null_fd)?;
+        dup2_stdout(&null_fd)?;
+        dup2_stderr(&null_fd)?;
+    }
+
+    unsafe {
+        let pid = libc::fork();
+        if pid < 0 {
+            bail!("fork error {}", std::io::Error::last_os_error());
+        } else if pid > 0 {
+            _exit(0);
+        }
+    }
 
     info!("emulating soft_reboot!");
     if let Err(e) = reset_boot_completed() {
